Fix windows warning

+++ test/unittest/types/endpoint/EndpointTest.cpp
@@ -23,7 +23,7 @@ using namespace eprosima::databroker;
 
 // Get a random QoS configuration
 QoS random_qos(
-        uint seed = 0)
+        uint16_t seed = 0)
 {
     DurabilityKind durability;
     ReliabilityKind reliability;
@@ -64,14 +64,14 @@ QoS random_qos(
 
 // Get a random topic name
 RealTopic random_topic(
-        uint seed = 0)
+        uint16_t seed = 0)
 {
     return RealTopic("TopicName_" + std::to_string(seed), "TopicType_" + std::to_string(seed));
 }
 
 // Get a random topic name
 EndpointKind random_endpoint_kind(
-        uint seed = 0)
+        uint16_t seed = 0)
 {
     if (seed % 2)
     {
@@ -85,7 +85,7 @@ EndpointKind random_endpoint_kind(
 
 // Get a random guid
 Guid random_valid_guid(
-        uint seed = 0)
+        uint16_t seed = 0)
 {
     eprosima::fastrtps::rtps::GuidPrefix_t guid_prefix;
     std::istringstream("44.53.00.5f.45.50.52.4f.53.49.4d." + std::to_string(seed)) >> guid_prefix;
@@ -160,7 +160,7 @@ TEST(EndpointTest, guid_getter)
 
     // Random guids
     {
-        for (uint i = 0; i < 10; i++)
+        for (uint16_t i = 0; i < 10; i++)
         {
             Guid guid = random_valid_guid(i);
             Endpoint endpoint(kind, guid, qos, topic);
@@ -183,7 +183,7 @@ TEST(EndpointTest, qos_getter)
 
     // Random guids
     {
-        for (uint i = 0; i < 8; i++)
+        for (uint16_t i = 0; i < 8; i++)
         {
             QoS qos = random_qos(i);
             Endpoint endpoint(kind, guid, qos, topic);
@@ -206,7 +206,7 @@ TEST(EndpointTest, topic_getter)
 
     // Random guids
     {
-        for (uint i = 0; i < 10; i++)
+        for (uint16_t i = 0; i < 10; i++)
         {
             RealTopic topic = random_topic(i);
             Endpoint endpoint(kind, guid, qos, topic);
